requirements.txt
disnake~=2.8.1
flask~=2.3.2
python-dotenv~=1.0.0